+++ src/main/java/com/example/timesaver/model/dto/applicants/selection/BulkUpdateSelectionRequest.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 @Data
 public class BulkUpdateSelectionRequest {
-    private List<Long> applicantIds;
+    private List<Integer> applicantIds;
     private Boolean selected; // set this status for all listed applicants
 }
