+++ src/main/java/com/example/timesaver/repository/AssignmentRepository.java
@@ -4,6 +4,6 @@ import com.example.timesaver.model.Assignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
-    List<Assignment> findByProjectProjectId(Long projectId);
+public interface AssignmentRepository extends JpaRepository<Assignment, Integer> {
+    List<Assignment> findByProjectProjectId(Integer projectId);
 }
