+++ src/main/java/com/example/timesaver/repository/QuestionAnswerRepository.java
@@ -3,12 +3,14 @@ package com.example.timesaver.repository;
 import com.example.timesaver.model.*;
 import com.example.timesaver.model.dto.applicants.display.QuestionAnswerDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.NativeQuery;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface QuestionAnswerRepository extends JpaRepository<QuestionAnswer, Long> {
+public interface QuestionAnswerRepository extends JpaRepository<QuestionAnswer, Integer> {
     @Query("SELECT new com.example.timesaver.model.dto.applicants.display.QuestionAnswerDTO(qa.question.questionNumber, qa.questionAnswer)  FROM QuestionAnswer qa WHERE qa.applicant.applicantId = :applicantId")
-    List<QuestionAnswerDTO> findQuestionNumberAndAnswerByApplicantId(@Param("applicantId") Long applicantId );
+    List<QuestionAnswerDTO> findQuestionNumberAndAnswerByApplicantId(@Param("applicantId") Integer applicantId );
+
 }
